separate clients route, add history and config
