+++ ui/app/controllers/vault/cluster/clients/config.js
@@ -1,0 +1,3 @@
+import Controller from '@ember/controller';
+
+export default class ConfigController extends Controller {}
