+++ ui/app/controllers/vault/cluster/clients/index.js
@@ -1,6 +1,3 @@
 import Controller from '@ember/controller';
 
-export default class ClientsController extends Controller {
-  queryParams = ['tab']; // ARG TODO remove
-  tab = null;
-}
+export default class ClientsController extends Controller {}
