+++ ui/app/styles/components/tabs.scss
@@ -44,3 +44,29 @@
     outline: none;
   }
 }
+
+.nav-tab-link {
+  color: $grey;
+  font-weight: $font-weight-semibold;
+  text-decoration: none;
+  padding: $size-6 $size-8 $size-8;
+  border-bottom: 2px solid transparent;
+  transition: background-color $speed, border-color $speed;
+
+  &:hover,
+  &:active {
+    border-color: $grey-light;
+  }
+
+  &:hover {
+    background-color: $ui-gray-050;
+  }
+
+  &:focus {
+    box-shadow: none;
+  }
+  &.is-active {
+    border-color: $blue !important;
+    color: $blue !important;
+  }
+}
